Keynote from ReactConf
React Native is a way of writing native mobile apps using the principles
of ReactJS. Pretty mind blowing to see what they've done.

This could be a really interesting development for us and could be a
serious consideration to allow us as a team to have more involvement in
the mobile apps.