60fps on the mobile web
The Flipboard engineering team explains how they used canvas as an
alternative to using the DOM so they could get 60fps scrolling on their
mobile web version.

Given the upcoming UX project, it may be interesting for us to begin to
forumlate ideas on how we would approach making the new UI work. In
particular, if we're scrolling large lists of images on mobile, the
techniques mentioned here may be relevant.