Rebuilding HipChat with React
An engineer from HipChat discusses their recent rewrite and why they
chose ReactJS.

Given that it seems we'll be doing more javascript in the future, it
would be good to have an understanding of when ReactJS would be an
appropriate choice. This is a good summary of the benefits in a real
world application.