One Language a Year - Elixir
http://devslovebacon.com/conferences/bacon-2014/talks/one-language-a-year-elixir

Learning new languages is good for the brain, even if we don't have a
use case for it right now.

This looks an interesting talk and although it's not immediately
applicable to us, as in we won't be writing any Elixir anytime soon,
I still think it would be a good video to watch.