The Product is the Byproduct
People in Silicon Valley perpetuate this idea that if you're smart enough, have an amazing vision, and keep pushing at it, then you'll end up with a great product. Take a look at startup success stories, though: a ton of them succeeded in spite of their original vision. A great product is closer to an accident. It's the byproduct of the environment you build at your company. This environment may actually be harder to build than the product itself, but you'll be left with a better everything by the end of it.