How GitHub Uses GitHub to Build GitHub
Build features fast. Ship them. That's what we try to do at GitHub. Our process is the anti-process: what's the minimum overhead we can put up with to keep our code quality high, all while building features *as quickly as possible*? It's not just features, either: faster development means happier developers.

We are working on the web, and our customers need our development to be as agile as possible so that CV as a company can experiment, innovate and evolve in its approach to our two primary visions, *touch a billion* and *impact a nation*.

Video: https://www.youtube.com/watch?v=qyz3jkOBbQY
Info: http://zachholman.com/talk/how-github-uses-github-to-build-github/